+++ bkmonitor/webpack/src/apm/pages/home/app-list.tsx
@@ -42,8 +42,8 @@ import { Debounce } from '../../../monitor-common/utils/utils';
 import EmptyStatus from '../../../monitor-pc/components/empty-status/empty-status';
 import { EmptyStatusOperationType, EmptyStatusType } from '../../../monitor-pc/components/empty-status/types';
 import GuidePage from '../../../monitor-pc/components/guide-page/guide-page';
-import { INodeType, TargetObjectType } from '../../../monitor-pc/components/monitor-ip-selector/typing';
-import { TimeRangeType } from '../../../monitor-pc/components/time-range/time-range';
+import type { INodeType, TargetObjectType } from '../../../monitor-pc/components/monitor-ip-selector/typing';
+import type { TimeRangeType } from '../../../monitor-pc/components/time-range/time-range';
 import { handleTransformToTimestamp } from '../../../monitor-pc/components/time-range/utils';
 // import DashboardTools from '../../../monitor-pc/pages/monitor-k8s/components/dashboard-tools';
 import AlarmTools from '../../../monitor-pc/pages/monitor-k8s/components/alarm-tools';
