+++ bkmonitor/webpack/src/apm/pages/service/service.tsx
@@ -28,7 +28,7 @@ import { Component, InjectReactive, Prop, Ref } from 'vue-property-decorator';
 import { Component as tsc } from 'vue-tsx-support';
 
 import { random } from '../../../monitor-common/utils/utils';
-import { TimeRangeType } from '../../../monitor-pc/components/time-range/time-range';
+import type { TimeRangeType } from '../../../monitor-pc/components/time-range/time-range';
 import { handleTransformToTimestamp } from '../../../monitor-pc/components/time-range/utils';
 import CommonNavBar from '../../../monitor-pc/pages/monitor-k8s/components/common-nav-bar';
 import CommonPage, { SceneType } from '../../../monitor-pc/pages/monitor-k8s/components/common-page-new';
