+++ bkmonitor/webpack/src/monitor-pc/components/temporary-share/history-share-manage.tsx
@@ -25,6 +25,7 @@
  */
 import { Component, Prop, Ref, Watch } from 'vue-property-decorator';
 import { Component as tsc } from 'vue-tsx-support';
+import { getDateValueToString } from '@blueking/date-picker/dist/vue2-full.es';
 import { Button, Checkbox, Table, TableColumn, TableSettingContent } from 'bk-magic-vue';
 import moment from 'moment';
 
@@ -358,7 +359,7 @@ export default class HistoryShareManage extends tsc<IProps> {
       if (!!timeRangeItem?.default_time_range) {
         timeRangeStr =
           this.shortcutsMap.get(timeRangeItem.default_time_range.join(' -- ')) ||
-          timeRangeItem.default_time_range.join(' -- ');
+          getDateValueToString(timeRangeItem?.default_time_range);
       } else {
         const formatStr = 'YYYY-MM-DD HH:mm:ss';
         timeRangeStr = `${moment(timeRangeItem.start_time * 1000).format(formatStr)}--${moment(
