+++ bkmonitor/webpack/src/monitor-pc/package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@blueking/bk-weweb": "0.0.22",
+    "@blueking/date-picker": "^0.0.11",
     "@blueking/ip-selector": "0.0.1-beta.125",
     "@blueking/paas-login": "0.0.6",
     "@blueking/user-selector": "^1.0.5-beta.2",
