+++ bkmonitor/webpack/src/monitor-pc/pages/data-retrieval/typings/index.ts
@@ -28,7 +28,7 @@ import { TranslateResult } from 'vue-i18n';
 
 import { random } from '../../../../monitor-common/utils/utils';
 import { EmptyStatusOperationType, EmptyStatusType } from '../../../components/empty-status/types';
-import { TimeRangeType } from '../../../components/time-range/time-range';
+import type { TimeRangeType } from '../../../components/time-range/time-range';
 import { PanelHeaderType, PanelToolsType } from '../../monitor-k8s/typings';
 import { IFunctionItem } from '../../strategy-config/strategy-config-set-new/monitor-data/function-menu';
 import { IFunctionsValue } from '../../strategy-config/strategy-config-set-new/monitor-data/function-select';
