+++ bkmonitor/webpack/src/monitor-pc/pages/monitor-k8s/components/common-page.tsx
@@ -45,7 +45,7 @@ import {
 import { VariablesService } from '../../../../monitor-ui/chart-plugins/utils/variable';
 import Collapse from '../../../components/collapse/collapse';
 import { ASIDE_COLLAPSE_HEIGHT } from '../../../components/resize-layout/resize-layout';
-import { TimeRangeType } from '../../../components/time-range/time-range';
+import type { TimeRangeType } from '../../../components/time-range/time-range';
 import { DEFAULT_TIME_RANGE } from '../../../components/time-range/utils';
 import { PANEL_INTERVAL_LIST } from '../../../constant/constant';
 import { Storage } from '../../../utils';
@@ -1167,8 +1167,8 @@ export default class CommonPage extends tsc<ICommonPageProps, ICommonPageEvent> 
         selectortTarget.compareFieldsSort
       );
       // eslint-disable-next-line max-len
-      compareTargets = targets?.map(item =>
-        selectortTarget?.handleCreateFilterDictValue(item, true, selectortTarget.compareFieldsSort)
+      compareTargets = targets?.map(
+        item => selectortTarget?.handleCreateFilterDictValue(item, true, selectortTarget.compareFieldsSort)
       );
     }
     const variables: Record<string, any> = {
