+++ bkmonitor/webpack/src/monitor-pc/pages/performance/performance-detail/compare-panel.vue
@@ -177,7 +177,6 @@
           </div>
           <div class="time-shift">
             <time-range
-              :type="showText ? 'normal' : 'simple'"
               :value="tools.timeRange"
               @change="handleTimeRangeChange"
             />
