+++ bkmonitor/webpack/src/monitor-pc/pages/view-detail/view-detail.vue
@@ -208,7 +208,7 @@ import MonitorDialog from '../../../monitor-ui/monitor-dialog/monitor-dialog.vue
 import MonitorEcharts from '../../../monitor-ui/monitor-echarts/monitor-echarts-new.vue';
 import { DEFAULT_REFLESH_LIST, DEFAULT_TIME_RANGE_LIST, DEFAULT_TIMESHIFT_LIST } from '../../common/constant';
 import SortButton from '../../components/sort-button/sort-button';
-import { TimeRangeType } from '../../components/time-range/time-range';
+import type { TimeRangeType } from '../../components/time-range/time-range';
 import { handleTransformToTimestamp } from '../../components/time-range/utils';
 // import { handleTimeRange } from '../../utils/index';
 import authorityMixinCreate from '../../mixins/authorityMixin';
