+++ bkmonitor/webpack/src/trace/components/time-range/time-range.tsx
@@ -24,24 +24,12 @@
  * IN THE SOFTWARE.
  */
 
-import { defineComponent, PropType, ref, toRefs, watch } from 'vue';
-import { Button, DatePicker, Input, Popover } from 'bkui-vue';
-import moment from 'moment';
+import { defineComponent, PropType } from 'vue';
+import DatePicker from '@blueking/date-picker/dist/vue3-full.es';
 
-import IconFont from '../icon-font/icon-font';
-
-import {
-  DEFAULT_TIME_RANGE,
-  handleTransformTime,
-  handleTransformToTimestamp,
-  intTimestampStr,
-  shortcuts,
-  shortcutsMap,
-  TimeRangeType
-} from './utils';
+import { TimeRangeType } from './utils';
 
 import './time-range.scss';
-
 /**
  * 图表选择时间范围组件
  */
@@ -54,227 +42,21 @@ export default defineComponent({
       default: () => ['now-1h', 'now']
     }
   },
-  setup(props, { emit }) {
-    const { modelValue } = toRefs(props);
-
-    /** 日历组件的值 */
-    const timestamp = ref<TimeRangeType>(['', '']);
-
-    /** 面板展示状态 */
-    const open = ref(false);
-    /** 是否展示面板的时间段 */
-    const isPanelTimeRange = ref(false);
-
-    /** 组件选中的时间范围 */
-    const localValue = ref<TimeRangeType>(['now-1h', 'now']);
-
-    const timeRangeDisplay = ref(shortcutsMap.get(DEFAULT_TIME_RANGE.join(' -- ')));
-    /**
-     * 控制日历面板是否展示
-     * @param show
-     */
-    const handleShowPanel = (show: boolean) => {
-      open.value = show ?? !open.value;
+  setup(p, { emit }) {
+    const handleChange = (value: TimeRangeType[]) => {
+      emit('update:modelValue', value);
     };
-    const handleOpenChange = (val: boolean) => {
-      !val && (open.value = val);
-    };
-    /**
-     * 处理trigger的时间展示
-     */
-    const timeRangeDisplayChange = () => {
-      const timeArr = isPanelTimeRange.value ? timestamp.value : localValue.value;
-      let timeDisplay = timeArr.join(' -- ');
-      if (shortcutsMap.get(timeDisplay)) {
-        timeDisplay = shortcutsMap.get(timeDisplay);
-      }
-      timeRangeDisplay.value = timeDisplay;
-    };
-
-    /** 将value转换成时间区间 */
-    const handleTransformTimeValue = (value: TimeRangeType = modelValue.value) => {
-      const dateArr = handleTransformTime(value);
-      timestamp.value = [dateArr[0], dateArr[1]];
-    };
-
-    /** 日历组件值变更 */
-    const handleDatePickerChagne = (date: TimeRangeType) => {
-      timestamp.value = date.map((item, index) => `${item} ${!index ? '00:00:00' : '23:59:59'}`) as TimeRangeType;
-      isPanelTimeRange.value = true;
-    };
-
-    /** 校验之间范围的合法性 */
-    const handleValidateTimeRange = (): boolean => {
-      const timeRange = handleTransformToTimestamp(modelValue.value);
-      /** 时间格式错误 */
-      if (timeRange.some(item => !item)) return false;
-      /** 时间范围有误 */
-      if (timeRange[0] > timeRange[1]) return false;
-      return true;
-    };
-    /** 将value转换成时间区间 */
-    // const handleTransformTimestamp = (value: TimeRangeType = modelValue.value) => {
-    //   const dateArr = handleTransformTime(value);
-    //   timestamp.value = [dateArr[0], dateArr[1]];
-    // };
-    /** 确认操作 */
-    const handleConfirm = () => {
-      const pass = handleValidateTimeRange();
-      if (!pass) {
-        localValue.value = [...modelValue.value];
-        open.value = false;
-      } else {
-        handleTimeRangeChange();
-        timeRangeDisplayChange();
-      }
-    };
-    /** 格式化绝对时间点 */
-    const formatTime = (value: TimeRangeType) =>
-      value.map(item => {
-        const m = moment(intTimestampStr(item));
-        return m.isValid() ? m.format('YYYY-MM-DD HH:mm:ss') : item;
-      });
-    /** 对外更新值 */
-    const handleTimeRangeChange = () => {
-      open.value = false;
-      handleTransformTimeValue(timestamp.value);
-      const newVal = isPanelTimeRange.value ? timestamp.value : formatTime(localValue.value);
-      emit('timeChange', newVal);
-      emit('update:modelValue', newVal);
-    };
-
-    /**
-     * 自定义输入时间范围
-     * @param index 值索引
-     */
-    const handleCustomInput = (index: number) => {
-      isPanelTimeRange.value = false;
-      if (!localValue.value[index]) {
-        localValue.value[index] = moment().format('YYYY-MM-DD HH:mm:ss');
-      }
-    };
-
-    /** 点击快捷时间选项 */
-    const handleShortcutChange = data => {
-      if (!!data?.value) {
-        isPanelTimeRange.value = false;
-        const value = [...data.value] as TimeRangeType;
-        handleTransformTimeValue(value);
-        localValue.value = value;
-      }
-      handleTimeRangeChange();
-      timeRangeDisplayChange();
-    };
-
-    watch(
-      () => props.modelValue,
-      val => {
-        localValue.value = val;
-        handleTransformTimeValue();
-        timeRangeDisplayChange();
-      }
-    );
-
     return {
-      open,
-      timestamp,
-      localValue,
-      timeRangeDisplay,
-      isPanelTimeRange,
-      handleShowPanel,
-      handleOpenChange,
-      handleDatePickerChagne,
-      handleConfirm,
-      handleCustomInput,
-      handleShortcutChange,
-      handleTransformTimeValue
+      handleChange
     };
   },
   render() {
     return (
       <DatePicker
         class='time-range-date-picker'
-        extPopoverCls='time-range-popover'
-        type={'daterange'}
-        value={this.timestamp}
-        open={this.open}
-        appendToBody
-        onOpen-change={this.handleOpenChange}
-        onChange={this.handleDatePickerChagne}
-        v-slots={{
-          trigger: () => (
-            <Popover
-              placement='bottom'
-              theme='light time-range-tips'
-              onAfterShow={() => {
-                this.handleTransformTimeValue();
-              }}
-              v-slots={{
-                content: () => (
-                  <div class='time-range-tips-content'>
-                    <div>{this.timestamp[0]}</div>
-                    <div>to</div>
-                    <div>{this.timestamp[1]}</div>
-                  </div>
-                )
-              }}
-            >
-              <span
-                class={['time-range-trigger', { active: this.open }]}
-                onClick={() => this.handleShowPanel(true)}
-              >
-                <IconFont
-                  icon='icon-mc-time'
-                  classes={['icon-time-range']}
-                ></IconFont>
-                <span>{this.timeRangeDisplay}</span>
-              </span>
-            </Popover>
-          ),
-          confirm: () => (
-            <div class='time-range-footer'>
-              <Button
-                theme='primary'
-                onClick={this.handleConfirm}
-              >
-                {this.$t('确定')}
-              </Button>
-            </div>
-          ),
-          header: () => (
-            <i18n-t
-              keypath='从 {0} 至 {1}'
-              tag='div'
-              class='time-range-custom'
-              // 20231025 禁止 MouseUp 事件冒泡是为了防止点击 header 插槽空间时导致整个弹窗关闭。（原因不明，暂时无法调试出是哪里的问题）
-              onMouseup={e => e.stopPropagation()}
-            >
-              <Input
-                class='custom-input'
-                v-model={this.localValue[0]}
-                onInput={() => this.handleCustomInput(0)}
-              />
-              <Input
-                class='custom-input'
-                v-model={this.localValue[1]}
-                onInput={() => this.handleCustomInput(1)}
-              />
-            </i18n-t>
-          ),
-          shortcuts: () => (
-            <ul class='shortcuts-list'>
-              {shortcuts.map(item => (
-                <li
-                  class='shortcuts-item'
-                  onClick={() => this.handleShortcutChange(item)}
-                >
-                  {item.text}
-                </li>
-              ))}
-            </ul>
-          )
-        }}
-      ></DatePicker>
+        modelValue={this.modelValue}
+        onUpdate:modelValue={this.handleChange}
+      />
     );
   }
 });
