+++ bkmonitor/webpack/src/trace/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@antv/g6": "^4.7.0",
+    "@blueking/date-picker": "^0.0.11",
     "@blueking/ip-selector": "0.0.1-beta.125",
     "bkui-vue": "0.0.2-beta.24",
     "d3-array": "^3.1.1",
@@ -35,7 +36,7 @@
     "reselect": "^4.1.5",
     "resize-detector": "^0.3.0",
     "throttle-debounce": "^3.0.1",
-    "vue": "3.2.33",
+    "vue": "^3.3.7",
     "vue-demi": "^0.14.6",
     "vue-i18n": "9.1.9",
     "vue-json-pretty": "2.2.2",
